Adds state animation to survive super power event
Enables custom battle events to react to state animation changes.

This allows for more dynamic and responsive events that can adapt to the current game state.
A condition property is added to manage and track state animation conditions.
