+++ src/js/custom-battle-events/survive-super-power-with-guard/procedures/create-survive-super-power-with-guard-props.ts
@@ -9,4 +9,5 @@ export const createSurviveSuperPowerWithGuardProps =
     state: {
       isIntroductionComplete: false,
     },
+    stateAnimationCondition: null,
   });
